qca-ssdk: add support for ipq60xx
IPQ60xx uses a different codename for SSDK, so lets pass the correct one
as otherwise SSDK asumes we are building for the old MIPS SoC-s.

swconfig and PTP features will be disabled on all of the targets SSDK
can support in OpenWrt, so move those to the rest of make flags.

Signed-off-by: Robert Marko <robimarko@gmail.com>