ipq807x: add node for required analog PSGMII
Now that we use the upstream qca807x PHY driver we need to better define
each PHY node. For the upstream driver it's also needed to define the
analog PSGMII PHY to apply required workarounds to correctly make it
work.

The analog PSGMII PHY is always placed 5 address after the address of
the first qca807x PHY in the system.

While at it also rename some PHY node name to the correct values.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>